// src/graphql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
const defaultOptions = {} as const;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Dimension: any;
  HexColor: any;
  Quality: any;
};
// ...
export type PostQueryVariables = Exact<{
  slug?: InputMaybe<Scalars['String']>;
}>;


export type PostQuery = { postsCollection?: { items: Array<{ title?: string | null } | null> } | null };


export const PostDocument = gql`
    query post($slug: String) {
  postsCollection(where: {slug: $slug}) {
    items {
      title
    }
  }
}
    `;